Move conference controllers to their own directory
